server/pyproject.toml
#
# PROJECT OVERVIEW
#

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend implementation for Golden Shoe webshop"
authors = ["Mikhail Ivanovskii"]

#
# DEPENDECIES SETTINGS
#

[tool.poetry.dependencies]
python = "^3.10.5"
Django = "^4.0.6"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
mypy = "^0.961"
pyright = "^1.1.261"

#
# BUILD SETTINGS
#

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#
# LINGTING SETTINGS
#

[tool.black]
line-length = 80

[tool.pyright]
pythonVersion = '^3.10.5'
reportMissingImports = true
reportMissingTypeStubs = false
disableOrganizeImports = true
pythonPlatform = 'Darwin'
typeCheckingMode = 'basic'
